Use uv action in test_hooks.yml, replace optional-dependencies with dependency-groups, extract test fixtures to files
- Replace actions/setup-python with astral-sh/setup-uv in test_hooks.yml
- Pin all actions to commit hashes (checkout@v6, setup-uv@v8.0.0)
- Use uv sync/run instead of pip install/python -m pytest
- Replace [project.optional-dependencies] with [dependency-groups] in pyproject.toml
- Extract all MINIMAL_* inline strings from tests/conftest.py to tests/examples/
- Simplify pdk_root fixture to use shutil.copytree

Agent-Logs-Url: https://github.com/nikosavola/pdk-ci-workflow/sessions/7c6c5ff8-b25c-4c3f-a044-70c0f3694085
