+++ pyproject.toml
@@ -26,7 +26,7 @@ check_precommit_config = "hooks.check_precommit_config:main"
 check_version_sync = "hooks.check_version_sync:main"
 check_template_drift = "hooks.check_template_drift:main"
 
-[project.optional-dependencies]
+[dependency-groups]
 test = [
     "pytest>=7.0",
     "hypothesis>=6.0",
