+++ tests/examples/.changelog.d/changelog_template.jinja
@@ -1,0 +1,1 @@
+{# template #}
