+++ tests/examples/.github/workflows/release-drafter.yml
@@ -1,0 +1,10 @@
+jobs:
+  update_release_draft:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: release-drafter/release-drafter@v5
+name: Release Drafter
+'on':
+  push:
+    branches:
+    - main
