+++ tests/examples/.github/workflows/test_code.yml
@@ -1,0 +1,14 @@
+jobs:
+  pre-commit:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v4
+    - uses: pre-commit/action@v3
+  test_code:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v4
+    - run: make test
+name: Test code
+'on':
+  push: null
