+++ tests/examples/.pre-commit-config.yaml
@@ -1,0 +1,28 @@
+repos:
+- hooks:
+  - id: end-of-file-fixer
+  - id: trailing-whitespace
+  repo: https://github.com/pre-commit/pre-commit-hooks
+  rev: v4.5.0
+- hooks:
+  - entry: ruff format
+    id: ruff-format
+    language: system
+    name: ruff-format
+    types:
+    - python
+  - entry: ruff check
+    id: ruff-lint
+    language: system
+    name: ruff-lint
+    types:
+    - python
+  repo: local
+- hooks:
+  - id: nbstripout
+  repo: https://github.com/mwouts/nbstripout
+  rev: '0.6.0'
+- hooks:
+  - id: codespell
+  repo: https://github.com/codespell-project/codespell
+  rev: v2.3.0
