+++ tests/examples/README.md
@@ -1,0 +1,2 @@
+# My PDK
+pip install my-pdk==0.1.0
