+++ tests/examples/my_pdk/__init__.py
@@ -1,0 +1,4 @@
+"""My PDK package."""
+
+__version__ = "0.1.0"
+__all__ = ["cells", "tech"]
