+++ tests/examples/my_pdk/cells/__init__.py
@@ -1,0 +1,1 @@
+from .waveguides import *
