+++ tests/examples/my_pdk/tech.py
@@ -1,0 +1,7 @@
+import gdsfactory as gf
+
+LAYER = gf.typings.Layer
+LAYER_STACK = gf.typings.LayerStack
+LAYER_VIEWS = gf.typings.LayerViews
+cross_sections = {}
+routing_strategies = {}
