+++ tests/examples/pyproject.toml
@@ -1,0 +1,67 @@
+[build-system]
+requires = ["setuptools>=68.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "my-pdk"
+version = "0.1.0"
+description = "A test PDK"
+requires-python = ">=3.9"
+readme = "README.md"
+license = {file = "LICENSE"}
+keywords = ["python", "photonics"]
+authors = [{name = "Test Author"}]
+dependencies = ["gdsfactory>=7.0"]
+classifiers = ["Development Status :: 3 - Alpha"]
+
+[dependency-groups]
+dev = ["pytest", "pytest-cov", "pytest_regressions", "pre-commit"]
+docs = ["jupytext", "jupyter-book"]
+
+[tool.ruff]
+fix = true
+
+[tool.ruff.lint]
+select = ["B", "C", "D", "E", "F", "I", "T10", "UP", "W"]
+ignore = ["E501", "B008", "C901", "B905", "C408"]
+
+[tool.ruff.lint.pydocstyle]
+convention = "google"
+
+[tool.codespell]
+ignore-words-list = "te,ba,fpr,ro,nd,donot,schem"
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
+addopts = "--tb=short"
+
+[tool.setuptools.package-data]
+"*" = ["*.csv", "*.yaml", "*.yml", "*.gds", "*.lyp", "*.oas", "*.lyt"]
+
+[tool.setuptools.packages.find]
+where = ["."]
+include = ["my_pdk"]
+
+[tool.tbump.version]
+current = "0.1.0"
+
+[[tool.tbump.file]]
+src = "pyproject.toml"
+
+[[tool.tbump.file]]
+src = "my_pdk/__init__.py"
+
+[[tool.tbump.file]]
+src = "README.md"
+
+[tool.tbump.git]
+message_template = "Bump to {new_version}"
+tag_template = "v{new_version}"
+
+[tool.mypy]
+strict = true
+
+[tool.towncrier]
+directory = ".changelog.d"
+filename = "CHANGELOG.md"
+template = ".changelog.d/changelog_template.jinja"
