+++ tests/examples/tests/test_pdk.py
@@ -1,0 +1,3 @@
+from my_pdk import tech
+def test_import(): assert tech
+def test_difftest(): pass  # difftest()
